Safer GO_CONFIG searching!
The big change here is in how we cascade up directories to find
GO_CONFIG. Previously we relied on Bash's string manipulation, but if in
the event we run into a path with a forward-slash in the directory name
the whole script will break. Now, we use `cd` to walk up the path,
avoiding this problem.

Because we're using `cd`, I also added a patherror function for if it
fails for some reason, at shellcheck's behest.

Another change is that I use `pwd -L` instead of `pwd -P` so we don't muss
up any symlinks in the directory tree.

Another change is that we're checking GOBIN changes after we source the
GO_CONFIG, so we can set GOBIN in the GO_CONFIG without it being
overwritten. At the same time, we make sure it IS set if no changes to
it are made in the GO_CONFIG.