Improve robustness and messages in Ctags hooks
Rewrite the run_tags.rb script entirely. More robust and more clear what is happening and better about handling what might go wrong. It also is more flexible when not run as a hook. Includes help documentation because I would always forget how it worked and have to read the code myself to figure out what was going on.